chore: move FirestoreDatabase fuzzer to normal approach
